FileAPI/tests/submissions/Mozilla/slice.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Blob slice</title>
<link rel=help href="http://dev.w3.org/2006/webapi/FileAPI/#slide-method-algo">
<link rel=author title="Saurabh Anand" href="mailto:saurabhanandiit@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/FileAPI/support/Blob.js"></script>
<div id="log"></div>
<script>
test_blob(function() {
  var blobTemp = new Blob(["PASS"]);
  return blobTemp.slice();
}, {
  expected: "PASS",
  type: "",
  desc: "no-argument Blob slice"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(-6);
}, {
  expected: "STRING",
  type: "",
  desc: "negative start parameter, relativeStart is positive"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(-12);
}, {
  expected: "PASSSTRING",
  type: "",
  desc: "negative start parameter, relativeStart is 0"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(4);
}, {
  expected: "STRING",
  type: "",
  desc: "positive start parameter, start is min"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSTRING"]);
  return blobTemp.slice(12);
}, {
  expected: "",
  type: "",
  desc: "positive start parameter, size is min"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(0, -6);
}, {
  expected: "PASS",
  type: "",
  desc: "negative end parameter, relativeEnd is positive"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(0, -12);
}, {
  expected: "",
  type: "",
  desc: "negative end parameter, relativeEnd is 0"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(0, 4);
}, {
  expected: "PASS",
  type: "",
  desc: "positive end parameter, end is min"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(0, 12);
}, {
  expected: "PASSSTRING",
  type: "",
  desc: "positive end parameter, size is min"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice();
}, {
  expected: "PASSSTRING",
  type: "",
  desc: "no contentType parameter, relativeContentType should be empty"
});

// Note the following bugs:
// https://www.w3.org/Bugs/Public/show_bug.cgi?id=18521
// https://www.w3.org/Bugs/Public/show_bug.cgi?id=18523
// https://www.w3.org/Bugs/Public/show_bug.cgi?id=18528
// https://www.w3.org/Bugs/Public/show_bug.cgi?id=18529

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "\xFF");
  });
}, "relativeContentType consists of non-ASCII characters, throw SyntaxError");

test_blob(function() {
  var blobTemp = new Blob(["PASS"]);
  return blobTemp.slice(0, 4, "TEXT/PLAIN");
}, {
  expected: "PASS",
  type: "text/plain",
  desc: "upper case contentType should be converted to lower case"
});

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te(xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('(')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te)xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (')')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te<xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('<')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te>xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('>')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te@xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('@')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te,xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (',')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te;xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (';')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te:xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (':')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\\xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('\\')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\"xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('\"')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te/xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('/')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te[xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('[')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te]xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (']')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te?xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('?')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te=xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('=')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te{xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('{')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te}xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('}')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\x20xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (SP)");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\x09xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (HT)");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\x00xt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError (0)");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\x1Fxt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('\x1F')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "te\x7Fxt/plain");
  });
}, "type consists of any CTLs or separators, throw SyntaxError ('\x7F')");

test(function() {
  var blobTemp = new Blob(["PASS"]);
  assert_throws("SyntaxError", function() {
    blobTemp.slice(0, 4, "text/plain;charset = UTF-8");
  });
}, "parameter attribute and value have LWS, throw SyntaxError");

test_blob(function() {
  var blobTemp = new Blob(["PASS"]);
  return blobTemp.slice(0, 4, "text/plain;charset=UTF-8")
}, {
  expected: "PASS",
  type: "text/plain;charset=UTF-8",
  desc: "contentType should be returned correctly"
});

test_blob(function() {
  var blobTemp = new Blob(["PASSSTRING"]);
  return blobTemp.slice(7, 4);
}, {
  expected: "",
  type: "",
  desc: "span is 0"
});
</script>
